+++ Sources/Views/ALKTemplateMessageCell.swift
@@ -8,15 +8,25 @@
 import UIKit
 
 open class ALKTemplateMessageCell: UICollectionViewCell {
+    // ALKTemplateMessageCell Can be customized via this.
+    public enum Style {
+        public static var borderColor : UIColor = UIColor.blue
+        public static var textColor: UIColor = UIColor.black
+        public static var cornorRadious: CGFloat = 10.0
+        public static var textSize: CGFloat = 16.0
+        public static var borderWidth: CGFloat = 1.0
+        public static var backgroundColor: UIColor = UIColor.clear
+    }
+    
     open var textLabel: UILabel = {
         let label = UILabel(frame: CGRect.zero)
-        label.textColor = UIColor.black
+        label.textColor = Style.textColor
         label.contentMode = .center
         label.numberOfLines = 1
-        label.font = Font.normal(size: 16.0).font()
+        label.font = Font.normal(size: Style.textSize).font()
         return label
     }()
-
+    
     public let leftPadding: CGFloat = 5.0
 
     override public init(frame: CGRect) {
@@ -31,10 +41,10 @@ open class ALKTemplateMessageCell: UICollectionViewCell {
 
     func setupUI() {
         layer.masksToBounds = true
-        layer.cornerRadius = 10.0
-        layer.borderWidth = 1
-        layer.borderColor = UIColor.gray.cgColor
-        backgroundColor = UIColor.clear
+        layer.cornerRadius = Style.cornorRadious
+        layer.borderWidth = Style.borderWidth
+        layer.borderColor = Style.borderColor.cgColor
+        backgroundColor = Style.backgroundColor
         addViewsForAutolayout(views: [textLabel])
 
         textLabel.topAnchor.constraint(equalTo: topAnchor).isActive = true